+++ changelog.md
@@ -1,3 +1,7 @@
 ### 2021-06-19
 - added: base model and xception model
 - [WIP] to create a model input pipeline
+
+### 2021-06-20
+- fixed: issues with the Xception model and removed base model arch
+- added: trainer and model manager for managing the models file